Moved functions to fit file structure, refactored affine cipher code to fit module format, upfated README with planned ciphers to implement
