+++ cryptotools/ciphers/__init__.py
@@ -1,0 +1,1 @@
+import affine_cipher
